alloy_json_rpc

Enum ResponsePayload

pub enum ResponsePayload<Payload = Box<RawValue>, ErrData = Box<RawValue>> {
    Success(Payload),
    Failure(ErrorPayload<ErrData>),
}

A JSON-RPC 2.0 response payload.

This enum covers both the success and error cases of a JSON-RPC 2.0 response. It is used to represent the result and error fields of a response object.

Note

This type does not implement Serialize or Deserialize directly. It is deserialized as part of the Response type.

Variants

Success(Payload)

A successful response payload.

Failure(ErrorPayload<ErrData>)

An error response payload.

Implementations

impl ResponsePayload<&RawValue, &RawValue>

pub fn into_owned(self) -> ResponsePayload

Convert this borrowed response payload into an owned payload by copying the data from the deserializer (if necessary).

impl<Payload, ErrData> ResponsePayload<Payload, ErrData>

pub const fn parse_error() -> Self

Create a new error payload for a parse error.

pub const fn invalid_request() -> Self

Create a new error payload for an invalid request.

pub const fn method_not_found() -> Self

Create a new error payload for a method not found error.

pub const fn invalid_params() -> Self

Create a new error payload for an invalid params error.

pub const fn internal_error() -> Self

Create a new error payload for an internal error.

pub const fn internal_error_message(message: Cow<'static, str>) -> Self

Create a new error payload for an internal error with a custom message.

pub const fn internal_error_with_obj(data: ErrData) -> Self
where ErrData: RpcSend,

Create a new error payload for an internal error with a custom message and additional data.

pub const fn internal_error_with_message_and_obj( message: Cow<'static, str>, data: ErrData, ) -> Self
where ErrData: RpcSend,

Create a new error payload for an internal error with a custom message and additional data.

pub const fn as_success(&self) -> Option<&Payload>

Fallible conversion to the successful payload.

pub const fn as_error(&self) -> Option<&ErrorPayload<ErrData>>

Fallible conversion to the error object.

pub fn try_into_success(self) -> Result<Payload, ErrorPayload<ErrData>>

Converts this type into a Result returning the ErrorPayload as Err.

pub fn error_code(&self) -> Option<i64>

Returns the error code if this a ResponsePayload::Failure

pub fn error_data(&self) -> Option<&ErrData>

Returns the error data if this a ResponsePayload::Failure

pub const fn is_success(&self) -> bool

Returns true if the response payload is a success.

pub const fn is_error(&self) -> bool

Returns true if the response payload is an error.

impl<Payload, ErrData> ResponsePayload<Payload, ErrData>
where Payload: RpcSend, ErrData: RpcSend,

pub fn serialize_payload(&self) -> Result<ResponsePayload>

Convert the inner types into a RawValue by serializing them.

impl<'a, Payload, ErrData> ResponsePayload<Payload, ErrData>
where Payload: AsRef<RawValue> + 'a,

pub fn try_success_as<T: Deserialize<'a>>(&'a self) -> Option<Result<T>>

Attempt to deserialize the success payload, borrowing from the payload if necessary.

Returns

None if the payload is an error

Some(Ok(T)) if the payload is a success and can be deserialized

Some(Err(serde_json::Error)) if the payload is a success and canโt be deserialized as T

pub fn deserialize_success<T: DeserializeOwned>( self, ) -> Result<ResponsePayload<T, ErrData>, Self>

Deserialize a Success payload, if possible, transforming this type.

Returns

Ok(ResponsePayload<T>) if the payload is an error, or if the payload is a success and can be deserialized as T

Err(self) if the payload is a success and canโt be deserialized

impl<'a, Payload, Data> ResponsePayload<Payload, Data>
where Data: Borrow<RawValue> + 'a,

pub fn try_error_as<T: Deserialize<'a>>(&'a self) -> Option<Result<T>>

Attempt to deserialize the error payload, borrowing from the payload if necessary.

Returns

None if the payload is a success

Some(Ok(T)) if the payload is an error and can be deserialized

Some(Err(serde_json::Error)) if the payload is an error and canโt be deserialized as T

pub fn deserialize_error<T: DeserializeOwned>( self, ) -> Result<ResponsePayload<Payload, T>, Self>

Deserialize an Error payload, if possible, transforming this type.

Returns

Ok(ResponsePayload<Payload, T>) if the payload is an error, or if the payload is an error and can be deserialized as T.

Err(self) if the payload is an error and canโt be deserialized.