Module alloy_json_rpc::packet

Enums

RequestPacket

A RequestPacket is a SerializedRequest or a batch of serialized request.

ResponsePacket

A ResponsePacket is a Response or a batch of responses.

ResponsePacketErrorsIter 🔒

An Iterator over the ErrorPayloads in a ResponsePacket.

Type Aliases

BorrowedResponsePacket

A BorrowedResponsePacket is a ResponsePacket that has been partially deserialized, borrowing its contents from the deserializer. This is used primarily for intermediate deserialization. Most users will not require it.