Struct alloy_json_rpc::request::SerializedRequest

pub struct SerializedRequest {
    meta: RequestMeta,
    request: Box<RawValue>,
}

A JSON-RPC 2.0 request object that has been serialized, with its Id and method preserved.

This struct is used to represent a request that has been serialized, but not yet sent. It is used by RPC clients to build batch requests and manage in-flight requests.

Fields
meta: RequestMetarequest: Box<RawValue>
Implementations

impl SerializedRequest

pub const fn meta(&self) -> &RequestMeta

Returns the request metadata (ID and Method).

pub const fn id(&self) -> &Id

Returns the request ID.

pub fn method(&self) -> &str

Returns the request method.

pub fn set_is_subscription(&mut self)

Mark the request as a non-standard subscription (i.e. not eth_subscribe)

pub fn is_subscription(&self) -> bool

Returns true if the request is a subscription.

pub const fn serialized(&self) -> &RawValue

Returns the serialized request.

pub fn into_serialized(self) -> Box<RawValue>

Consume the serialized request, returning the underlying RawValue.

pub fn decompose(self) -> (RequestMeta, Box<RawValue>)

Consumes the serialized request, returning the underlying RequestMeta and the RawValue.

pub fn take_request(self) -> Box<RawValue>

Take the serialized request, consuming the SerializedRequest.

pub fn params(&self) -> Option<&RawValue>

Get a reference to the serialized request’s params.

This partially deserializes the request, and should be avoided if possible.

pub fn params_hash(&self) -> B256

Get the hash of the serialized request’s params.

This partially deserializes the request, and should be avoided if possible.