alloy_json_rpc

Trait BorrowedRpcObject

pub trait BorrowedRpcObject<'de>: RpcBorrow<'de> + RpcSend { }

An object that can be both sent and received over RPC, borrowing from the the deserialization context.

This marker trait is blanket-implemented for every qualifying type. It is used to indicate that a type can be both sent and received in the body of a JSON-RPC message, and can borrow from the deserialization context.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.