alloy_json_rpc

Type Alias BorrowedResponsePacket

pub type BorrowedResponsePacket<'a> = ResponsePacket<&'a RawValue, &'a RawValue>;

A BorrowedResponsePacket is a ResponsePacket that has been partially deserialized, borrowing its contents from the deserializer.

This is used primarily for intermediate deserialization. Most users will not require it.

See the top-level docs for more info.

Aliased Type

enum BorrowedResponsePacket<'a> {
    Single(Response<&'a RawValue, &'a RawValue>),
    Batch(Vec<Response<&'a RawValue, &'a RawValue>>),
}

Variants

Single(Response<&'a RawValue, &'a RawValue>)

A single response.

Batch(Vec<Response<&'a RawValue, &'a RawValue>>)

A batch of responses.

Implementations

impl BorrowedResponsePacket<'_>

pub fn into_owned(self) -> ResponsePacket

Convert this borrowed response packet into an owned packet by copying the data from the deserializer (if necessary).