alloy_json_rpc

Type Alias PartiallySerializedRequest

pub type PartiallySerializedRequest = Request<Box<RawValue>>;

A Request that has been partially serialized.

The request parameters have been serialized, and are represented as a boxed RawValue. This is useful for collections containing many requests, as it erases the Param type. It can be created with Request::box_params().

See the top-level docs for more info.

Aliased Type

struct PartiallySerializedRequest {
    pub meta: RequestMeta,
    pub params: Box<RawValue>,
}

Fields
meta: RequestMeta

The request metadata (ID and method).

params: Box<RawValue>

The request parameters.